# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewAcceptedResponse": {
        "description": "\u5f02\u6b65\u5ba1\u67e5\u4efb\u52a1\u5165\u961f\u540e\u7684\u7a33\u5b9a\u786e\u8ba4\u54cd\u5e94\u3002",
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "review_run_id": {
            "title": "Review Run Id",
            "type": "string"
          },
          "review_task_id": {
            "title": "Review Task Id",
            "type": "string"
          },
          "review_version_key": {
            "title": "Review Version Key",
            "type": "string"
          }
        },
        "required": [
          "review_run_id",
          "review_task_id",
          "review_version_key",
          "execution_status",
          "accepted_at",
          "created"
        ],
        "title": "ReviewAcceptedResponse",
        "type": "object"
      },
...
      "ReviewActionRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ReviewAction"
          },
          "target_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewActionRequest",
        "type": "object"
      },
      "ReviewActionResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ReviewAction"
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "review_run_id": {
            "title": "Review Run Id",
            "type": "string"
          },
          "review_task_id": {
            "title": "Review Task Id",
            "type": "string"
          },
          "workflow_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action",
          "review_run_id",
          "review_task_id",
          "execution_status"
        ],
        "title": "ReviewActionResponse",
        "type": "object"
      },
...
      "ReviewChangeTokenResponse": {
        "properties": {
          "change_token": {
            "title": "Change Token",
            "type": "string"
          }
        },
        "required": [
          "change_token"
        ],
        "title": "ReviewChangeTokenResponse",
        "type": "object"
      },
...
      "ReviewFindingDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/FindingDecision"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewFindingDecisionRequest",
        "type": "object"
      },
...
      "ReviewListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items",
          "next_cursor"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
...
      "ReviewRequest": {
        "additionalProperties": false,
        "description": "\u4e00\u6b21\u5f02\u6b65 Pull Request \u5ba1\u67e5\u4f7f\u7528\u7684\u5185\u90e8\u8bf7\u6c42\u3002",
        "properties": {
          "head_sha": {
            "maxLength": 64,
            "minLength": 40,
            "title": "Head Sha",
            "type": "string"
          },
          "installation_id": {
            "exclusiveMinimum": 0.0,
            "title": "Installation Id",
            "type": "integer"
          },
          "pull_request_number": {
            "exclusiveMinimum": 0.0,
            "title": "Pull Request Number",
            "type": "integer"
          },
          "repository": {
            "maxLength": 255,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$",
            "title": "Repository",
            "type": "string"
          },
          "repository_id": {
            "exclusiveMinimum": 0.0,
            "title": "Repository Id",
            "type": "integer"
          }
        },
        "required": [
          "installation_id",
          "repository_id",
          "repository",
          "pull_request_number",
          "head_sha"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenReviewer API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "description": "\u8fd4\u56de\u8ba4\u8bc1\u540e\u7684\u6700\u8fd1\u5ba1\u67e5\u4efb\u52a1\u5217\u8868\u3002\n\n\u53c2\u6570\uff1a\n    _: \u7ba1\u7406\u5458\u4f1a\u8bdd\u4f9d\u8d56\u3002\n    limit: \u8fd4\u56de\u6761\u6570\uff0c\u8303\u56f4 1 \u5230 100\uff0c\u9ed8\u8ba4 50\u3002\n\n\u8fd4\u56de\uff1a\n    \u5305\u542b\u6570\u636e\u5e93\u4e2d\u7684\u603b\u8fd0\u884c\u6570\u548c\u6309\u521b\u5efa\u65f6\u95f4\u5012\u5e8f\u6392\u5217\u7684\u6700\u8fd1\u4efb\u52a1\uff1b\u4efb\u52a1\u72b6\u6001\u3001\n    \u5c1d\u8bd5\u6b21\u6570\u548c\u6700\u540e\u9519\u8bef\u6765\u81ea\u540c\u4e00\u6b21 Dashboard \u8bfb\u53d6\u3002\n\n\u5f02\u5e38\uff1a\n    HTTPException(401): \u4f1a\u8bdd\u65e0\u6548\u3002\n    HTTPException(503): \u67e5\u8be2\u5931\u8d25\u3002",
        "operationId": "list_reviews_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews"
      },
      "post": {
        "description": "\u6821\u9a8c\u5e42\u7b49\u952e\u5e76\u63a5\u53d7\u4e00\u4e2a\u5f02\u6b65\u5ba1\u67e5\u4efb\u52a1\u3002\n\n\u8bf7\u6c42\u5fc5\u987b\u5148\u901a\u8fc7\u4f1a\u8bdd\u548c\u540c\u6e90\u68c0\u67e5\uff1b\u4ed3\u50a8\u5c42\u4fdd\u8bc1\u8fd0\u884c\u3001\u4efb\u52a1\u548c Outbox \u4e8b\u4ef6\u5728\u540c\n\u4e00\u4e2a\u4e8b\u52a1\u4e2d\u6301\u4e45\u5316\u3002\u76f8\u540c\u952e\u91cd\u590d\u63d0\u4ea4\u8fd4\u56de\u539f\u4efb\u52a1\uff0c\u4e0d\u540c\u5185\u5bb9\u590d\u7528\u540c\u4e00\u952e\u5219\u8fd4\u56de\n409\uff0c\u6570\u636e\u5e93\u6682\u65f6\u4e0d\u53ef\u7528\u5219\u8fd4\u56de\u53ef\u5b89\u5168\u91cd\u8bd5\u7684 503\u3002\n\n\u53c2\u6570\uff1a\n    request_body: \u5df2\u901a\u8fc7 Pydantic \u4e25\u683c\u5b57\u6bb5\u6821\u9a8c\u7684\u5ba1\u67e5\u8bf7\u6c42\u3002\n    idempotency_key: HTTP ``Idempotency-Key``\uff0c\u957f\u5ea6 1 \u5230 200\uff1b\u51fd\u6570\u4f1a\u518d\u53bb\u6389\n        \u9996\u5c3e\u7a7a\u767d\uff0c\u7a7a\u767d\u952e\u8fd4\u56de 422\u3002\n    _: \u7ba1\u7406\u5458\u4f1a\u8bdd\u4f9d\u8d56\u7ed3\u679c\uff0c\u4ec5\u7528\u4e8e\u786e\u8ba4\u8c03\u7528\u65b9\u5df2\u767b\u5f55\u3002\n    __: \u540c\u6e90\u4f9d\u8d56\u7ed3\u679c\uff0c\u4ec5\u7528\u4e8e\u963b\u6b62\u5e26\u6076\u610f Origin \u7684\u6d4f\u89c8\u5668\u526f\u4f5c\u7528\u8bf7\u6c42\u3002\n\n\u8fd4\u56de\uff1a\n    202 \u54cd\u5e94\uff0c\u5305\u542b\u8fd0\u884c/\u4efb\u52a1 ID\u3001\u7248\u672c\u952e\u3001\u5f53\u524d\u6267\u884c\u72b6\u6001\u3001\u9996\u6b21\u63a5\u53d7\u65f6\u95f4\u548c\n    ``created`` \u6807\u5fd7\u3002\u91cd\u590d\u8bf7\u6c42\u7684\u72b6\u6001\u53ef\u80fd\u5df2\u7ecf\u4ece ``queued`` \u63a8\u8fdb\u5230\u5176\u4ed6\u72b6\u6001\u3002\n\n\u5f02\u5e38\uff1a\n    HTTPException(401/403/422/409/503): \u5206\u522b\u5bf9\u5e94\u4f1a\u8bdd\u65e0\u6548\u3001Origin \u4e0d\u5339\u914d\u3001\n    \u8bf7\u6c42\u6216\u5e42\u7b49\u952e\u4e0d\u5408\u6cd5\u3001\u952e\u6307\u5411\u4e0d\u540c\u5185\u5bb9\u3001\u6301\u4e45\u5316\u4e0d\u53ef\u7528\u3002",
        "operationId": "create_review_api_v1_reviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review"
      }
    },
...
    "/api/v1/reviews/{review_run_id}": {
      "get": {
        "description": "\u8fd4\u56de\u4efb\u52a1\u7684\u9636\u6bb5\u3001\u6a21\u578b\u7ed3\u679c\u3001Finding\u3001CI \u548c\u7ed3\u6784\u5316\u4e8b\u4ef6\u65e5\u5fd7\u3002",
        "operationId": "get_review_details_api_v1_reviews__review_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "finding_limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Finding Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "finding_cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Finding Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review Details"
      }
    },
    "/api/v1/reviews/{review_run_id}/actions": {
      "post": {
        "description": "\u6267\u884c\u53ef\u5ba1\u8ba1\u7684\u52a0\u901f\u3001\u91cd\u8bd5\u3001\u53d6\u6d88\u6216\u91cd\u65b0\u5ba1\u67e5\u52a8\u4f5c\u3002",
        "operationId": "apply_review_action_api_v1_reviews__review_run_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Review Action"
      }
    },
    "/api/v1/reviews/{review_run_id}/change-token": {
      "get": {
        "description": "\u7528\u4e00\u6b21\u7d22\u5f15\u67e5\u8be2\u8fd4\u56de\u8be6\u60c5\u53d8\u5316\u4ee4\u724c\uff0c\u4e0d\u52a0\u8f7d\u5b50\u8d44\u6e90\u3002",
        "operationId": "get_review_change_token_api_v1_reviews__review_run_id__change_token_get",
        "parameters": [
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewChangeTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review Change Token"
      }
    },
    "/api/v1/reviews/{review_run_id}/findings/{finding_id}": {
      "post": {
        "description": "\u4fdd\u5b58 Finding \u7684\u201c\u786e\u8ba4\u95ee\u9898/\u5ffd\u7565\u201d\u88c1\u51b3\u5e76\u8fd4\u56de\u6700\u65b0\u8be6\u60c5\u3002",
        "operationId": "decide_review_finding_api_v1_reviews__review_run_id__findings__finding_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewFindingDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Review Finding"
      }
    },
    "/api/v1/reviews/{review_run_id}/identity/sync": {
      "post": {
        "description": "\u4ece GitHub \u56de\u67e5\u5e76\u8865\u5168\u5386\u53f2\u4efb\u52a1\u7684 PR \u4f5c\u8005\u3001\u94fe\u63a5\u548c\u5206\u652f\u4fe1\u606f\u3002",
        "operationId": "sync_review_identity_api_v1_reviews__review_run_id__identity_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Review Identity"
      }
    },
...
  }
}